requirements.txt.txt
flask
flask-cors
numpy
joblib
pillow
opencv-python
tensorflow==2.13.0
scikit-learn==1.3.2
